remove userComponentVisible hook and add cross icon to make qr modal invisible
